pyproject.toml
[project]
name = "lm-forge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=2.18.0",
    "huggingface-hub>=0.20.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "safetensors>=0.4.0",
    "torch>=2.0.0",
    "transformers>=4.40.0",
]
